Remove GLFW and GLEW from Utility
