+++ src/Engine/CMakeLists.txt
@@ -121,7 +121,7 @@ set_property(SOURCE ${SRCS} APPEND PROPERTY OBJECT_DEPENDS ${EMBEDDED_HEADER})
 create_directory_groups(${SRCS} ${HEADERS})
 
 add_library(Engine STATIC ${SRCS} ${HEADERS} ${EMBEDDED_SOURCE})
-target_link_libraries(Engine Video Physics catch imgui stb simpleini ${OPENAL_LIB} angelscript)
+target_link_libraries(Engine Video Physics catch glfw ${GLFW_LIBRARIES} imgui stb simpleini ${OPENAL_LIB} angelscript)
 
 # Measure VRAM dependencies.
 if(MeasureVRAM)
