+++ src/Engine/MainWindow.cpp
@@ -5,10 +5,13 @@
 #include "Manager/Managers.hpp"
 #include "Manager/RenderManager.hpp"
 #include "Util/Input.hpp"
-#include <Utility/Log.hpp>
 
 MainWindow* MainWindow::instance = nullptr;
 void WindowSizeCallback(GLFWwindow* window, int width, int height);
+void ErrorCallback(int error, const char* description);
+void HandleDebugMessage(GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei length, const GLchar* message, const void* userParam, bool showNotifications);
+void APIENTRY DebugMessageCallback(GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei length, const GLchar* message, const void* userParam);
+void APIENTRY DebugMessageCallbackIgnoreNotifications(GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei length, const GLchar* message, const void* userParam);
 
 MainWindow::MainWindow(int width, int height, bool fullscreen, bool borderless, const char* title, bool debugContext) {
     
@@ -108,4 +111,93 @@ void WindowSizeCallback(GLFWwindow* window, int width, int height) {
 		MainWindow::GetInstance()->SetSize(width, height);
 		Managers().renderManager->UpdateBufferSize();
 	}
+}
+
+void ErrorCallback(int error, const char* description) {
+    fputs(description, stderr);
+}
+
+void HandleDebugMessage(GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei length, const GLchar* message, const void* userParam, bool showNotifications) {
+    if (!showNotifications && severity == GL_DEBUG_SEVERITY_NOTIFICATION)
+        return;
+    
+    switch (source) {
+    case GL_DEBUG_SOURCE_API:
+        fputs("Open GL API", stderr);
+        break;
+    case GL_DEBUG_SOURCE_WINDOW_SYSTEM:
+        fputs("Window System", stderr);
+        break;
+    case GL_DEBUG_SOURCE_SHADER_COMPILER:
+        fputs("Shader Compiler", stderr);
+        break;
+    case GL_DEBUG_SOURCE_APPLICATION:
+        fputs("Application", stderr);
+        break;
+    default:
+        fputs("Other", stderr);
+    }
+    
+    fputs(": ", stderr);
+    
+    switch (type) {
+    case GL_DEBUG_TYPE_ERROR:
+        fputs("Error", stderr);
+        break;
+    case GL_DEBUG_TYPE_DEPRECATED_BEHAVIOR:
+        fputs("Deprecated Behavior", stderr);
+        break;
+    case GL_DEBUG_TYPE_UNDEFINED_BEHAVIOR:
+        fputs("Undefined Behavior", stderr);
+        break;
+    case GL_DEBUG_TYPE_PORTABILITY:
+        fputs("Portability", stderr);
+        break;
+    case GL_DEBUG_TYPE_PERFORMANCE:
+        fputs("Performance", stderr);
+        break;
+    case GL_DEBUG_TYPE_MARKER:
+        fputs("Marker", stderr);
+        break;
+    case GL_DEBUG_TYPE_PUSH_GROUP:
+        fputs("Push Group", stderr);
+        break;
+    case GL_DEBUG_TYPE_POP_GROUP:
+        fputs("Pop Group", stderr);
+        break;
+    default:
+        fputs("Other", stderr);
+    }
+    
+    fputs(" (", stderr);
+    
+    switch (severity) {
+    case GL_DEBUG_SEVERITY_HIGH:
+        fputs("High Priority", stderr);
+        break;
+    case GL_DEBUG_SEVERITY_MEDIUM:
+        fputs("Medium Priority", stderr);
+        break;
+    case GL_DEBUG_SEVERITY_LOW:
+        fputs("Low Priority", stderr);
+        break;
+    case GL_DEBUG_SEVERITY_NOTIFICATION:
+        fputs("Notification", stderr);
+        break;
+    }
+    
+    fputs("):\n", stderr);
+    
+    fputs(message, stderr);
+    fputs("\n\n", stderr);
+    
+    fflush(stderr);
+}
+
+void APIENTRY DebugMessageCallback(GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei length, const GLchar* message, const void* userParam) {
+    HandleDebugMessage(source, type, id, severity, length, message, userParam, true);
+}
+
+void APIENTRY DebugMessageCallbackIgnoreNotifications(GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei length, const GLchar* message, const void* userParam) {
+    HandleDebugMessage(source, type, id, severity, length, message, userParam, false);
 }
